Tests: Relocated restart related tests from mainTest to restartTest
Added a restart BDE test
